Refactor profile progress to ts
