+++ packages/web/src/components/profile-progress/ProfileCompletionPanel.module.css
@@ -1,0 +1,43 @@
+.container {
+  background-color: var(--harmony-n-50);
+  border: 1px solid var(--harmony-n-150);
+  height: 96px;
+  width: 198px;
+  border-radius: 8px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+}
+
+.outerPadding {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 5px;
+  margin-bottom: 22px;
+  width: 198px;
+}
+
+.completionText {
+  color: var(--harmony-n-600);
+  font-size: 12px;
+  font-weight: bold;
+  line-height: 15px;
+  letter-spacing: 0.43px;
+}
+
+.dismissButton {
+  background-color: transparent;
+  border: none;
+  font-size: 12px;
+  font-weight: 500;
+  letter-spacing: 0.43px;
+  line-height: 15px;
+  color: var(--harmony-n-500);
+  cursor: pointer;
+}
+
+.dismissButton:hover {
+  text-decoration: underline;
+}
