+++ packages/web/src/components/profile-progress/components/ProfileCompletionTooltip.tsx
@@ -1,17 +1,23 @@
+import { ReactNode } from 'react'
+
 import Tooltip from 'components/tooltip/Tooltip'
 
 import { getPercentageComplete } from './ProfileCompletionHeroCard'
 import styles from './ProfileCompletionTooltip.module.css'
-import { CompletionStageArray } from './PropTypes'
-import TaskCompletionList from './TaskCompletionList'
+import { TaskCompletionList } from './TaskCompletionList'
+import { CompletionStages } from './types'
 
-const makeStrings = ({ completionPercentage }) => ({
+const makeStrings = (completionPercentage: number) => ({
   completionPercentage: `Profile ${completionPercentage}% Complete`
 })
 
-const TooltipContent = ({ completionStages }) => {
+type TooltipContentProps = {
+  completionStages: CompletionStages
+}
+
+const TooltipContent = ({ completionStages }: TooltipContentProps) => {
   const completionPercentage = getPercentageComplete(completionStages).toFixed()
-  const strings = makeStrings({ completionPercentage })
+  const strings = makeStrings(Number(completionPercentage))
 
   return (
     <div className={styles.content}>
@@ -24,17 +30,23 @@ const TooltipContent = ({ completionStages }) => {
   )
 }
 
+type ProfileCompletionTooltipProps = {
+  completionStages: CompletionStages
+  children: ReactNode
+  isDisabled?: boolean
+  shouldDismissOnClick?: boolean
+}
+
 /**
  * ProfileCompletionTooltip is a hovering tooltip that presents the
  * percentage of profile completion and the list of completion stages.
- *
- * @param {Object} { completionStages }
  */
-const ProfileCompletionTooltip = ({
+export const ProfileCompletionTooltip = ({
   completionStages,
   children,
-  isDisabled
-}) => {
+  isDisabled,
+  shouldDismissOnClick = false
+}: ProfileCompletionTooltipProps) => {
   return (
     <Tooltip
       color='secondary'
@@ -42,7 +54,7 @@ const ProfileCompletionTooltip = ({
       className={styles.tooltip}
       disabled={isDisabled}
       mouseEnterDelay={0.1}
-      shouldDismissOnClick={false}
+      shouldDismissOnClick={shouldDismissOnClick}
       text={<TooltipContent completionStages={completionStages} />}
       placement='right'
       mount={null}
@@ -51,9 +63,3 @@ const ProfileCompletionTooltip = ({
     </Tooltip>
   )
 }
-
-ProfileCompletionTooltip.propTypes = {
-  completionStages: CompletionStageArray.isRequired
-}
-
-export default ProfileCompletionTooltip
