--- packages/web/src/components/profile-progress/components/PropTypes.js
@@ -1,8 +1,0 @@
-import PropTypes from 'prop-types'
-
-export const CompletionStage = PropTypes.shape({
-  title: PropTypes.string.isRequired,
-  isCompleted: PropTypes.bool.isRequired
-})
-
-export const CompletionStageArray = PropTypes.arrayOf(CompletionStage)
