+++ packages/web/src/components/profile-progress/components/TaskCompletionList.tsx
@@ -1,25 +1,29 @@
 import cn from 'classnames'
-import PropTypes from 'prop-types'
 
-import { CompletionStageArray } from './PropTypes'
-import TaskCompletionItem from './TaskCompletionItem'
+import { TaskCompletionItem } from './TaskCompletionItem'
 import styles from './TaskCompletionList.module.css'
+import { CompletionStages } from './types'
 
-const sortIncompleteFirst = (list) => {
+const sortIncompleteFirst = (list: CompletionStages) => {
   const incomplete = list.filter((e) => !e.isCompleted)
   const complete = list.filter((e) => e.isCompleted)
   return incomplete.concat(complete)
+}
+
+type TaskCompletionListProps = {
+  completionStages: CompletionStages
+  className?: string
 }
 
 /**
  * `TaskCompletionList` renders a list of `TaskCompletionItems`.
  * It's used to power the lists inside of `ProfileCompletionTooltip`
  * and `ProfileCompletionHeroCard`.
- *
- * @param {Object} { completionStages, className }
- * @returns
  */
-const TaskCompletionList = ({ completionStages, className }) => (
+export const TaskCompletionList = ({
+  completionStages,
+  className
+}: TaskCompletionListProps) => (
   <div className={cn(styles.container, className)}>
     {sortIncompleteFirst(completionStages).map((e) => (
       <TaskCompletionItem
@@ -30,11 +34,3 @@ const TaskCompletionList = ({ completionStages, className }) => (
     ))}
   </div>
 )
-
-TaskCompletionList.propTypes = {
-  //  The order of stages in the array is determines the order of stages displayed.
-  completionStages: CompletionStageArray.isRequired,
-  className: PropTypes.string
-}
-
-export default TaskCompletionList
