+++ packages/web/src/components/profile-progress/components/types.ts
@@ -1,0 +1,6 @@
+export type CompletionStage = {
+  title: string
+  isCompleted: boolean
+}
+
+export type CompletionStages = CompletionStage[]
